fix: Images not appearing in page content.
I believe that the issue was related to the github subdomain, which confused the build and created
the image URL's incorrectly.
There is still work todo, as the content needs to be brought up to current, for the latest version etc..
However, I would like to get some user testing and feedback, and also organise the hosting, which I plan to
do way from github.